refactor: :art: 重建项目架构
+ 加入模型层
+ 加入辅助函数层
+ 重构部分代码适应新的项目架构
--沈嘉乐
