+++ app/services/auth.py
@@ -1,6 +1,5 @@
-from app.services.mongo import User
-from app.config import Config
-import re
+from app.utils.utils import check_password_hash, generate_password_hash
+from app.models.user import User
 import secrets
 class UserService:
     @staticmethod
@@ -34,25 +33,3 @@ class UserService:
     @staticmethod
     def create_token() -> str:
         return secrets.token_urlsafe(32)
-
-def generate_password_hash(password: str) -> str:
-    return Config.pwd_context.hash(password)
-
-
-def check_password_hash(password: str, hash: str) -> bool:
-    return Config.pwd_context.verify(password, hash)
-
-
-def is_strong_password(password):
-    """检查密码强度"""
-    if len(password) < 8:
-        return False
-    if not re.search(r"[A-Z]", password):
-        return False
-    if not re.search(r"[a-z]", password):
-        return False
-    if not re.search(r"[0-9]", password):
-        return False
-    if not re.search(r"[!@#$%^&*()_+=-]", password):
-        return False
-    return True
